Done some code refactoring. Database is improved (4 columns - watched, favorite, custom, watchlist - have been replaced with one column MovieList). Fixed issues caused by database changes.

+++ app/Http/Controllers/MovieController.php
@@ -2,14 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use App\MovieList;
 use Illuminate\Http\Request;
 use App\Movie;
-use App\Favorite;
-use App\Custom;
-use App\Watchlist;
 use App\Acting;
 use App\Cast;
-use App\Watched;
 use App\Comment;
 use App\User;
 use App\Rating;
@@ -27,48 +24,14 @@ class MovieController extends Controller
     {
         //get all movies ordered by latest added and return movie index
         $movies = Movie::latest()->get();
-        return $this->movieViewDetails($movies);       
+        return $this->movieViewDetails($movies);
     }
 //---------------------------------------------------------------------------------\\
 
 //---------------------------------SHOW MOVIE--------------------------------------\\
     public function show(Movie $movie)
     {
-        //find all ratings for current movie, count ratings and find average rating
-        $ratings = Rating::where('movie_id', 'LIKE', $movie->id)->get();
-        $ratingNum = (int) $ratings->count();
-        $ratingSum=0;
-        foreach($ratings as $rating) {
-            $ratingSum=$ratingSum+(int) $rating->rating;
-        }
-        
-        if($ratingNum==0) {
-            $rating = 0;
-        } else {
-            $rating = round($ratingSum/$ratingNum);
-        }
-        
-        //find all details which are needed to show on movie show page
-        $actings = Acting::where('movie_id', '=', $movie->id)->get();
-        $stars = Acting::where('movie_id', '=', $movie->id)->latest()->take(3)->get();
-        $director = Cast::where('occupation', '=', 'director')->first();
-        $sumWatched = Watched::where('movie_id', '=', $movie->id)->count();
-        $comments = Comment::where('movie_id', '=', $movie->id)->withLikes()->latest()->paginate(5);
-        $sumComments = Comment::where('movie_id', '=', $movie->id)->count();
-
-        //return movie show page with 
-        return view('movies.show',[
-            'movie'=>$movie,
-            'actings'=>$actings,
-            'director'=>$director,
-            'sumWatched'=>$sumWatched,
-            'comments'=>$comments,
-            'stars'=>$stars,
-            'sumComments'=>$sumComments,
-            'rating' => $rating,
-            'ratingNum' => $ratingNum,
-            'ratings' => $ratings
-        ]);
+       return $this->showMovie($movie);
     }
 //----------------------------------------------------------------------------------\\
 
@@ -78,7 +41,7 @@ class MovieController extends Controller
         //authorize so only admin can see add new movie button
         $this->authorize('isAdmin', Auth::user());
 
-        //get all actors 
+        //get all actors
         $actors = Cast::all();
 
         //return to view that show form for adding new movie
@@ -105,7 +68,7 @@ class MovieController extends Controller
         ]);
 
         $movie = new Movie();
-        
+
         $movie->name=request('name');
         $movie->genre=request('genre');
         $movie->releaseYear=request('releaseYear');
@@ -141,7 +104,7 @@ class MovieController extends Controller
             $acting->cast_id=$id;
             $acting->save();
         }
-    
+
         //return to the add new movie page
         return back();
     }
@@ -169,7 +132,7 @@ class MovieController extends Controller
             'descrption'=>'required',
         ]);
 
-        $movie->name=$name;
+        $movie->name=request('name');
         $movie->genre=request('genre');
         $movie->releaseYear=request('releaseYear');
         $movie->imdb=request('imdb');
@@ -183,7 +146,7 @@ class MovieController extends Controller
             Image::make($avatar)->save(public_path('/images/movies/' . $filename));
             $movie->avatar = $filename;
         }
-        
+
         $movie->save();
 
         //return back to the edit movie form
@@ -192,7 +155,7 @@ class MovieController extends Controller
 //----------------------------------------------------------------------------------\\
 
 //--------------------------------SORT MOVIES---------------------------------------\\
-    public function sortMovies(Request $request) 
+    public function sortMovies(Request $request)
     {
         //get value of sort
         if(isset($_GET['sort'])) {
@@ -200,7 +163,7 @@ class MovieController extends Controller
         } else {
             $sort = '';
         }
-        
+
         //sort users from sort value
         switch ($sort) {
             case 'ratingHighest':
@@ -212,104 +175,82 @@ class MovieController extends Controller
             case 'newest':
                 $movies = Movie::orderBy('created_at', 'desc')->get();
                 break;
-            case 'oldest': 
+            case 'oldest':
                 $movies = Movie::orderBy('created_at')->get();
                 break;
-            case 'favoriteMost':
-                $movies = Movie::withCount('favorite')->orderBy('favorite_count', 'desc')->get();
-                break;
-            case 'favoriteLeast': 
-                $movies = Movie::withCount('favorite')->orderBy('favorite_count')->get();
-                break;
-            case 'watchedMost': 
-                $movies = Movie::withCount('watched')->orderBy('watched_count', 'desc')->get();
-                break;
-            case 'watchedLeast': 
-                $movies = Movie::withCount('watched')->orderBy('watched_count')->get();
-                break;
-            case 'commentsMost': 
+            case 'commentsMost':
                 $movies = Movie::withCount('comment')->orderBy('comment_count', 'desc')->get();
                 break;
-            case 'commentsLeast': 
+            case 'commentsLeast':
                 $movies = Movie::withCount('comment')->orderBy('comment_count')->get();
                 break;
-            case 'watched':
-                break;
-            default: 
+            default:
                 $movies = Movie::latest()->get();
 
         }
-        
+
         //return to the movie index with sorted movies
-        return $this->movieViewDetails($movies);       
+        return $this->movieViewDetails($movies);
     }
 //----------------------------------------------------------------------------------\\
 
 //-----------------------STORE AND DELETE FROM WISHED LIST--------------------------\\
-    public function storeListItem(Request $request, $addToList, Movie $movie) 
+    public function storeListItem(Request $request, $addToList, Movie $movie, $type)
     {
         //request route to know which operation is asked for
         $route = $request->path();
 
-        //add movie to the wished list and flash message 
+        //add movie to the wished list and flash message
         switch ($route) {
-            case 'addFavorite/'.$movie->id:
-                $favorite = Favorite::create(['user_id' => Auth::user()->id, 'movie_id' => $movie->id]);
+            case 'addFavorite/'.$movie->id.'/favorite':
+                $favorite = MovieList::create(['user_id' => Auth::user()->id, 'movie_id' => $movie->id, 'type' => $type]);
                 flash('Movie added to favorites!')->success();
                 break;
-            case 'addCustom/'.$movie->id:
-                $custom = Custom::create(['user_id' => Auth::user()->id, 'movie_id' => $movie->id]);
+            case 'addCustom/'.$movie->id.'/custom':
+                $custom = MovieList::create(['user_id' => Auth::user()->id, 'movie_id' => $movie->id, 'type' => $type]);
                 flash('Movie added to custom!')->success();
                 break;
-            case 'addWatchlist/'.$movie->id:
-                $watchlist = Watchlist::create(['user_id' => Auth::user()->id, 'movie_id' => $movie->id]);
+            case 'addWatchlist/'.$movie->id.'/watchlist':
+                $watchlist = MovieList::create(['user_id' => Auth::user()->id, 'movie_id' => $movie->id, 'type' => $type]);
                 flash('Movie added to watchlist!')->success();
                 break;
-            case 'addWatched/'.$movie->id: 
-                $watched = Watched::create(['user_id' => Auth::user()->id, 'movie_id' => $movie->id]);
-                
-                //if movie is marked as watched, then movie will be deleted from watchlist
-                $watchlist = Watchlist::where('user_id', '=', Auth::user()->id)->where('movie_id', '=', $movie->id);
-
-                if($watchlist!=null) {
-                    $watchlist->delete();
-                }
-
+            case 'addWatched/'.$movie->id.'/watched':
+                $watched = MovieList::create(['user_id' => Auth::user()->id, 'movie_id' => $movie->id, 'type' => $type]);
                 flash('Movie added to watched!')->success();
                 break;
         }
-        
+
         //return to the movie page
-        return back();
+        return $this->showMovie($movie);
     }
 
-    public function destroyFromList(Request $request, $destroyFromList, Movie $movie) 
+    public function destroyFromList(Request $request, $destroyFromList, Movie $movie, $type)
     {
         //request route to know which operation is asked for
         $route = $request->path();
 
-        //delete movie from the wished list and flash message 
+        //delete movie from the wished list and flash message
         switch ($route) {
-            case 'favoriteDestroy/'.$movie->id:
-                $favorite = Favorite::where('user_id', '=', Auth::user()->id)->where('movie_id', '=', $movie->id)->delete();
+            case 'favoriteDestroy/'.$movie->id.'/favorite':
+                $favorite = MovieList::where('user_id', '=', Auth::user()->id)->where('movie_id', '=', $movie->id)->where('type', '=', $type)->delete();
                 flash('Movie removed from favorites!')->success();
                 break;
-            case 'customDestroy/'.$movie->id:
-                $custom = Custom::where('user_id', '=', Auth::user()->id)->where('movie_id', '=', $movie->id)->delete();
+            case 'customDestroy/'.$movie->id.'/custom':
+                $custom = MovieList::where('user_id', '=', Auth::user()->id)->where('movie_id', '=', $movie->id)->where('type', '=', $type)->delete();
                 flash('Movie removed from custom!')->success();
                 break;
-            case 'watchlistDestroy/'.$movie->id:
-                $watchlist = Watchlist::where('user_id', '=', Auth::user()->id)->where('movie_id', '=', $movie->id)->delete();
+            case 'watchlistDestroy/'.$movie->id.'/watchlist':
+                $watchlist = MovieList::where('user_id', '=', Auth::user()->id)->where('movie_id', '=', $movie->id)->where('type', '=', $type)->delete();
                 flash('Movie removed from watchlist!')->success();
                 break;
-            case 'watchedDestroy/'.$movie->id: 
-                $watched = Watched::where('user_id', '=', Auth::user()->id)->where('movie_id', '=', $movie->id)->delete();
+            case 'watchedDestroy/'.$movie->id.'/watched':
+                $watched = MovieList::where('user_id', '=', Auth::user()->id)->where('movie_id', '=', $movie->id)->where('type', '=', $type)->delete();
                 flash('Movie removed from watched!')->success();
                 break;
         }
-        
+
         //return to the movie page
-        return back();
+        return $this->showMovie($movie);
     }
 //----------------------------------------------------------------------------------\\
 
@@ -320,5 +261,5 @@ class MovieController extends Controller
         return back();
     }
 //----------------------------------------------------------------------------------\\
- 
+
 }
